Safari-609.1.11.1/LayoutTests/fast/files/filereader-zip-bundle-expected.txt

Test that bundles are automatically zipped when accessed via FileReader. To test manually, please drag a file bundle to a "Choose File" input below.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS file.size > 400 is true
PASS reader.result is 'PK'
PASS successfullyParsed is true

TEST COMPLETE
